Use lock to prevent orphaned processes
Put the work that Worker does inside a lock statement, and
acquire the same lock object when stopping the service.
This will let the worker finish it's current work before the srvice
is stopped. Without this there is a possibility that the Worker
will leave orphaned umrol.exe processes when the service is stopped
while the Worker is in the middle of "working" and waiting for
a umrol.exe process to exit.